do not remove setup properties when reusing

+++ utils/_context/_scenarios/core.py
@@ -126,7 +126,7 @@ class Scenario:
 
         path = os.path.join(self.host_log_folder, subfolder)
 
-        if remove_if_exists:
+        if remove_if_exists and not os.environ.get('ST_REUSE'):
             shutil.rmtree(path, ignore_errors=True)
 
         Path(path).mkdir(parents=True, exist_ok=True)
